+++ src/types/Types.ts
@@ -1,0 +1,8 @@
+export type ValuesImgType = {
+  images: string[];
+  step: number;
+  frameSize: number;
+  itemWidth: number;
+  animationDuration: number;
+  infinite: boolean;
+};
